fixes #185.
This sets the default to 0 so that transpose is only
enabled at a configurable threshold if you want to use
it. I am not going to add support for wide tables
we haven't defined the use cases yet and too much is
in flight right now. It should be discussed further.