change angular UI's bind-html-unsafe to angular ng-bind-html
We are currently using old version of angular UI. The current version
has deprecated this api. Besides bind-html-unsafe ignores texts that
don't have any HTML tags. There are some cases that eventhough we trust
the value to be HTML, there might not be any HTML tags in it. For
instance if the row_markdown_pattern is just a simple string without any
HTML tags.