Removed 1 time binding
In `recordedit`, when the table directive is initialized for a foreign key popup and then later reused for the recordedit resultset page when multiple records are created, it was showing an actions heading for the actions column when it shouldn't be there. The one way binding was causing this conditional to not re-evaluate when the directive was reused.