remove tuples from reference.delete function
In ermrestjs we're not using the tuples, and also passing tuples
required some extra requests for getting the tuples.
In case of association table, the rows that we are deleting are
different from the ones that we show to users. Therefore we need
to do another extra read for each individual tuple to get the actual
tuple. These extra reads were occurring on load and was causing some
hiccups.