fix url limit check after submitting facets
For some issues that we faced in the past, I had to introduce an
attribute in chaise that keeps track of applied filters. Since it
was manually added in chaise, we had to make sure that it is always
aligned with the reference facets that ermrestjs is returning.
In case of submitting the facets, I was changing that variable first
and then was checking the reference for URL limit. So in that case
the reference and appliedFilters would be inconsistent and recordset
would break.