make sure record and recordset are using their own flowControlObject
Previously all the variables related to flow control logic was
attached to $rootScope and they were sharing the same variables.
Now I am attaching those variabels in an object in table-model.
Since it's an object all the related/inline entites in the record
app can share the same object and modify the same variables.