removing href=# from navbar dropdown links
This was causing issues in some browsers. Bootstrap is supposed to
stop the propagation of click event after opening the dropdown, so
the link that we're passing should be ignored. But in some browsers,
after opening the dropdown, it's going to the link that is defined
in the href which is "#". This will cause a redirect to a page that
chaise doesn't understand (because the hash which has the catalog and
table information is empty).

This will remove those link definitions.