fix check presence bug
The list of filters that we're showing to the users is different from
the actual list of filters that are used for faceting. We have done this
because we wanted to keep track of these fitlers. For the check presence,
instead of showing the null filter, we were always showing not-null filter.