remove tableModel extra attribute and check in place
I realized adding an attribute to tableModel will impact other places
of the code. This change is currently only needed for record page and
therefore it's better if it stays there.
Also we are planning on changing how we're showing the inline display
in near future and that can only be done in record page and not table.js