fix spacing of record and recordset
to make sure the top and bottom panels have the same size.
When there's a scrollbar in the bottom panel, the padding of the panel
needs to be adjusted. This has been done by adding a watch event
that looks for the difference between the width of top and bottom
panels and adds a padding to the bottom panel if needed.