fix race condition in afterReadAggregate
After the main read in recordset table is done, we're generating
requests for aggregate columns based on the active list. Also if we
have to show more than 500 cells, we're going to gradually populate
`rowValues` object. Now if we recieve the response for aggregate columns
before populating the object, it will result in a terminal error.

To fix this, I introduced a new object called `pendingRowValues`.
The purpose of this object is hold onto the row values while we're
waiting for the rowValues to be completely populated.
I also added a new function that the "push more rows" will call upon
finishing to merge the two objects.