change how we set main-container height and padding
Previously we relied on angularjs watchers to do these changes, but we
found some issues with this and that's why I changed them in this commit.
The issues were:

- Since we were watching for the change in height and width of main-container,
  which was running on each digest cycle, it was very inefficient.
- We realized that the changes to template were applied after calling the watchers.
  So if the template was going to change because of an angularjs change, the
  watcher would catch it at all until the second digest cycle. So in some cases,
  the height wouldn't even change. This was causing issues mainly in the popups.
- We were adjusting the padding on each digest cycle, which would easily cause
 terminal error.

Now I'm using a pure javascript library to watch for the element size changes, and
will run the padding and height logic only on change. This will ensure that we're
always setting the correct padding and height for the main-container. Although
since we're not doing this as frequently as before, you can notice the padding
change. To make this a little less noticable, I removed the initial padding for
the main-container.

While doing this I also realized the change that we did in resizable was buggy.
We used to pass rOtherElementSelector to define the other elements that should
resize with the main element. But in case of modals, this would cause the left panel
of both record app and the recordset modal to be in sync with each other.
To resolve this, I introduced a new rOtherElements scope variable so we can
pass actual elements. This way in recordset I can define the parent element
that the top-left-panel should be selected from.