use the new generateRelatedList api
We realized that the old API had a bug that would cache the result
regardless of the given output. This was causing a race condition
is record page between the two functions:

- If export annotation is missing, ermrestjs will call .related() during
the default export annotation generation.

- After the main read we will call .related(tuple) to generate the
list of related references that we should show.

If the first function gets called first, .related would not use
the tuple for generating the references which would mean not using
facets for url generartion. Which eventually would cause "Explore"
links to not show any filter chiclet in the recordset page.