fix range picker bug related to clearing facets
The function that was responsible for syncing the checkboxes was not
properly supporting the not-null choice. It was assuming that the
not-null choice is always available, and also wouldn't clear the
not-null filter at all. I modified the function to properly take
care of the not-null and the rest of the options.