fix a bug in ellipsis logic for aggregates
When aggregate values of a specific table instance returns,
we should run the ellipsis logic. In some pages, we
show multiple table instances and therefore we had to make sure
the emitted message was unique For each instance.
But the message that we were using was not unique enough.
It wasn’t supporting inline vs. related vs. popups and would
cause terminal error if both related and popup had aggregates.

With this change, I changed the message to use a unique uuid
for each of the table instances.