use upload directive instead of disabled input for select-all asset
a while ago we added disabled support to upload directive. so instead
using a simple text input for disabled assets, we should have used that
directive instead.

With this change, the select-all asset inputs are not relying on
submissionRows anymore and can function properly like a normal
disabled asset input.