remove trackby to rerender ellipsis on every data update
We were using `reference-modified` to manually rerender the ellipsis logic.
This was problematic because this event was in some cases called in the
middle of two digest cycles and before actually vm.rowValues was properly
populated. So I had to change how this was done.

At first I used the old method of watching for the `rowValues` in ellipsis.
This solved the issue but I realized by removing track by I can ensure
AngularJS will run the link function for us.

we were trying to manually rerendering the logic of ellipsis anyways. this
will just allow angularjs to properly do this for us. The only downside
would be destroying and creating new DOM. But I think it's the more
robust solution.