make sure vm.page and vm.rowValues are changed in same digest cycle + tests
While manually testing it I realized that in some cases we were showing
The stale data. I added more info to the code that explains why the 
vm.page instantiation should be in that place of the code.

I also added test cases to make sure we’re testing “search” in 
Facet and add pure and binary popups.