fix a bug where external links were not working
Since our addClickListener is looking for the element that is "closest"
to the target element, the target element and the element that we are
looking for might be different.

For instance, assume the case that an image is wrapped in a link. Clicking
on the image was triggering the click listener, but the returned `e.target`
was an image (and not a link that the code was assuming). So I changed the code
to return the desired target as well, and changed the callbacks to use the
new variable.