Fix the implementation of markdown preview to work in other apps
The code was naively clicking on a button in recordedit page to show
the markdown preview. I modified it to directly call the showModal
Function that we have. This way it can be used in other places as well.

While doing this, I also refactored the code a bit and removed other 
unused packages (angular-datepicker, rzslider).