Fix the default/preselected foreign key value in inputSwitch
We have two modes of using inputSwitch:
- select-all: doesn’t have initial value and the 
  Value that is stored and displayed is not necessarily 
  What will be sent and we’re using an extra translation layer.
- normal field (used in viewer): can have initial value and
  The stored value is exactly what will be sent.

By sent I don’t mean sent to ermrestjs. No matter what we’re still
going to have a translation layer that is called populateSubmissionRows.

So the problem with select-all case is that it has an EXTRA translation
layer.


So in the inputs.js I had to make sure both cases are supported but 
eventually we should modify select-all to be aligned with the normal case.