fix the order of setting page height and showing iframe
There are three different events that happen that should be ordered properly:

- (1) When all the requests are done and iframe location is updated, so we
  tell viewer app that its "displayReady".
- (2) When we initially set the height of the bottom-container.
- (3) When we show the iframe.

With the previous approach, we were doing (3) as soon as the HTML page loads.
Then (1) and (2) together. Since we were doing (1) and (2) together, by the time
that (2) was done, the page wasn't still showing anything and therefore it couldn't
set a proper height. So the iframe would show up with a small height. If the iframe
had to load a lot of images, the lag would delay the logic to update the height and
user could see the small iframe for a few seconds.

I modified the code to do (1) and in the second digest cycle it would do (2)
and (3) together. So by the time that we're setting the initial height the rest of the
page is properly displayed and the inital height is what we want. So you won't see a
small iframe on load.