change default markdown margin bottom for p, ul, etc
changed it from 16 to 8
in the cases that there's just a single p and ul,
the space between them were too much, so we reduced it.
we couldn't reduce this any further since in the cases of multiple p,
they would be too close each other and it was hard to read.