properly manage the query parameters that will be sent to osd viewer
There are three different sources for such parameters:

1. query parameters on the page
2. image.uri value
3. Image_channel rows (for channel info)

We will merge (1) and (2) together while priotizing (1) over (2) (if
there's a parameter in both, (1) will be used).

If there wasn't any non-annotation "url" attribute in (1) and (2), we
will get the channel info from (3).

It's worth mentioning that if there's a non-annotation "url" in any
of them, then the all the parameters related to channel will be derived
only from that source. For example if (1) had only channelName while
(2) had "url", we will ignore the channelName in (1).