viewer app: use before and after for fetch-by-zindex
Using lt and gt doesn't work properly with pagination.
if you say lt a number and ask for the first 5, you would have to
then change the sort to get the correct 5 values, otherwise you
will get rows that wasn't expected. using proper before and after
was the easier solution