viewer app: fix a race condition in loading annotations
when users click on a image in zplane, we have to update the displayed
image as well as the annotations on the image. To update the annotations,
- the image must be loaded
- chaise request to fetch the annotations should be done
since I wasn't checking for the image being loaded, sometimes the
annotations couldn't properly load. so with this I made sure both
conditions are met before changing the annotations