populate export options when the dropdown is enabled
the recordset generateFacetColumns logic is changing the Reference
location on load. it will ignore the facets that have issues and therefore
the export can function afterwards.

But before this change, export was looking at the facets and was trying
to parse them and therefore was throwing a separate error.

with this change, I made sure to populate the export options after the
button is enabled.