remove .chaise-body from chaise-button rules
the only reason to wrap these rules in chaise-body is to ensure
they don't apply to static sites and don't clash with their rules.

but the name is very specific to chaise and that shouldn't be an issue.
besides, because of this, the GO to RID button is not showing proper
color and border.