fix the resize icon in the resizable panels
when I changed the fontawesome from pro to free, I didn't realize that
the font that we're using here is only available in pro version.

I added a section to dev-guide to explain why that happened.

To summarize, we were using the css class for the solid version which is
available in free version. But using css rules we modified the font-weight
to use the light version which is not available in free version.