include fontawesome using webfonts instead of svg core
since we're allowing data-modelers to add classes to their markdown,
then can use the font awesome icons. while with svg method we're just
going to include the fonts that chaise is using..

later we might want to improve this by making sure only the fonts
that are used in both places are included, instead of including
everything.