handle empty input for colors like any other input
I assumed that I shouldn't allow selection of "empty value" in the case
of nullok=false columns. but this would make the input to show black
while users haven't selected anything.

so now the input will be empty just like any other inputs and angularjs
will properly complain if users try to submit the empty form if the
column is required.