rework the tooltips in selected-chiclet names
my previous solution to force showing the arrow on left doesn't work
whent the element was at the edge of the page. so instead I changed
the ellipsis logic to make sure the name width doesn't go beyond
the limit and therefore the tooltip arrowhead always shows up
in the middle of the name.