fix flow-control issues
In all the callbacks in flow-control functions,
we used the state variables and props, assuming their latest value is getting used.

But after a request is done, we call the passed updatePageCB callback.
This means that the updatePageCB will use the closure of the caller and
not necessarily the latest state of variables.

That's why I changed the props and state variables in flow-control-related
callbacks to use the proper "reference" instead.