fix the implementation main and related spinner + handle toc corner cases
the previous implementation was tied to the record provider logic and
was separate from the recordset provider. As a result,
there would've been cases where the recordsetState.page is undefined
while the spinner is set to true. So I changed  the implementation
to be based on what's displayed on the page and recordsetState.

This allowed me to properly implement the scroll-to and toc disabling logic