properly url-encode cookie values in CookieService
while fully url-encoding is not required, we have to make sure certain
characeters are not part of the value. Doing the full url-encode is more
straightforward and seems to be the convention.
(the angularjs's putObject is doing url encode/decode)

In the prefill logic of "add records" we're storing objects in cookie.
These raw and unencoded cookie values were causing issues in our
backend services that access the cookies (as whitepsace and double
quotes are not allowed in cookie value).