properly populate foreignKeyData in create mode
in create mode, if some columns have default or prefilled values,
we should send extra requests to make sure those are valid and
display proper rowname to users.

in AngularJS version, we didn't properly block users while those
requests are pending. in this commit, I made sure to show an overlay
on top of only the foreignkeys that use domain-filter (and also disabled
the clone button while waiting for requests).