fix the misalignment issue in recordedit cells
since we were using rows={5} for textareas, the default height would
result in 107.992px and the whole cell would be 126.992px.

In the resizeSensor for ensuring consistent height between entity-key
and cells, I was mistakenly using "offsetHeight" which would return
a rounded integer instead. so in this case, the height of entity-key
would be 126px and that's why you could see the misalingment issue.

Changing the resizeSensor to use `getBoundingClientRect` fixed the issue.
I also changed the min-height of textarea to be 110px, so the default
height is not set to a number with fractions.