fix the navbar z-index issue + clean up code
a while ago I realized that the default bootstrap z-index is overriding our custom
values for modals. But that change broke the z-index definition of all the modals.

in this commit, I moved all the z-index definitions of modal into one place. this way
we can be sure that the same rule/order applies to all the modals, and it's easier
to follow the code.