fix an issue related to help app and defaultTable logic
ConfigService is calling getCatalogId which in turn will call the
chaiseURItoErmrestURI function to see if there's any catalog encoded
in the url.

this function will turn query parameters into hash since we want to
be able to handle this for chaise urls, but as a result, apps that
expect query parameter will be treated as if its the hash fragment.

the code then uses this query parameter to find the catalog id and
if it cannot find the schema/table, it would consult the defaultTable
chaise config. in cases where the hash was actually a query parameter,
we should not even attempt the defaulTable logic.