refactor copy to clipboard functions + change tooltip as a feedback to users (#2365)
the share and cite popup links were not using the same function as permalink,
and were only using the deprecated API for clipboard which wasn't working
properly on some browsers.

we also decided to change the tooltip as a feedback to users. to make sure
the tooltip is displayed in the proper place I had to change the tooltip
component a bit.