use latest version of fontawesome
more info: https://github.com/informatics-isi-edu/chaise/issues/2386

this also includes a fix to alignment of chaise-btn-group. in firefox, the
align-item: baseline was causing issues. it couldn't properly align
items with different height. so while the content was properly aligned,
the bounding box of each button would not align with the rest.

Now we only add this rule for the cases where we're not showing any borders.
currently only used in one place but I decided to use a general name in case
we want to do something similar later.