use latest version of fontawesome (#2390)
more info: https://github.com/informatics-isi-edu/chaise/issues/2386

This also includes a fix to the alignment of .chaise-btn-group. 

In Firefox, the "align-item": baseline was causing issues as it couldn't 
correctly align items with different heights. So, while the content was 
properly aligned, the bounding box of each button would not align with the rest. 

Now, we only add this rule for the cases where we're not showing any borders 
(currently only used in one place, but I decided to use a general name in case 
we want to do something similar later).