remove unnecessary background color from narrow dropdown menus
With this rule the items in the submenu were not visible since by
default (and in all our static sites) we're using dark background color
and grey/white text color.

We now let the default boostrap or the value defined by the static sites
to be used (similar to how we're not defining any color for dropdown-menu
outside of this media query).