revert the change to dropdowns for fixing safari bug
The rule by itself was more harmful. because of this,
the dropdowns would go below the content if the parent didn't specify
any z-index (e.g., export button)

A simple fix for this is adding z-index:1 to all of our dropdowns
by default, but I wasn't sure if this was the best solution, so I
I just reverted it for now.