add test cases for facet-within-facet
test cases are not finalized yet. while the structure of test spec is finalized, I need to change the configuration.
And for that I need to first figure out why the data that I created is not working on some cases. I'll fix this in the
next set of commits.