fix a small issue in facet choice + force timezone in playwright
I refactored the functions in facet-choice-pciker.tsx to use promises,
but it seems like I forgot a line. This was to ensure we're skipping
the request when 10 or more options are already selected.

also the timezone change shouldn't affect anything and is just to
make sure we're always using the same timezone regardless of the system
that is running the tests.