adjust the file type detection in file-preview
The previous implementation used the file extension only if the content type was missing. But now we’re going to use the file extension if we don’t recognize the content type. This way, more cases will now show the file preview. All of our parsers are sanitizing the file content, so there isn’t any security issues with doing this.

So, for example, if you had an `example.cif` file with the generic `Content-type: application/octet-stream`, we were not showing the file preview. But now the file preview will show up.